+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -10,7 +10,6 @@ import datetime
 import numpy as np
 import scipy as sp
 import scipy.stats
-import warnings
 import sys
 from copy import deepcopy
 from itertools import combinations