+++ docs/sources/CHANGELOG.md
@@ -5,6 +5,27 @@
 The CHANGELOG for the current development version is available at
 [https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md](https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md).
 
+### Version 0.8.1dev (TBD)
+
+
+##### Downloads
+
+- [Source code (zip)](https://github.com/rasbt/mlxtend/archive/v0.8.1.zip)
+- [Source code (tar.gz)](https://github.com/rasbt/mlxtend/archive/v0.8.1.tar.gz)
+
+##### New Features
+
+- 
+
+##### Changes
+
+- 
+
+##### Bug Fixes
+
+- 
+
+
 ### Version 0.8.0 (2017-09-09)
 
 
@@ -83,9 +104,6 @@ The CHANGELOG for the current development version is available at
 
 ### Version 0.5.1 (2017-02-14)
 
-
-The CHANGELOG for the current development version is available at
-[https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md](https://github.com/rasbt/mlxtend/blob/master/docs/sources/CHANGELOG.md).
 
 ##### Downloads